+++ coinpaprika/src/endpoint/globalMarketcap.ts
