+++ coinpaprika/src/endpoint/index.ts
@@ -1,4 +1,4 @@
 export * as price from './price'
 export * as dominance from './dominance'
-export * as marketcap from './marketcap'
+export * as globalMarketcap from './globalMarketcap'
 export * as multi from './multi'
